+++ unity/EditorPlugin/EditorPluginUnity56.csproj
@@ -25,7 +25,7 @@
   <ItemGroup>
     <Compile Remove="AfterUnity56\UnitTesting\CompiledAssembliesTracker.cs" />
   </ItemGroup>
-    <Import Project="$(DotNetSdkPath)\Build\PackageReference.JetBrains.RdFramework.Props"/>
+    <Import Project="$(DotNetSdkPath)\Build\PackageReference.JetBrains.RdFramework.Props" />
   <Import Project="ILRepack.targets" />
   <Import Project="Model.targets" />
 </Project>
